Made pull request changes + a little code refactoring
