+++ components/QuizCardPage/Timeline/index.test.js
@@ -1,15 +1,8 @@
 import { render, screen } from "@testing-library/react";
 import Timeline from ".";
 
-test("renders Timeline component without crashing", () => {
-  render(
-    <Timeline
-      isTimerPaused={false}
-      setIsDisabled={() => {}}
-      timeLeft={30}
-      setTimeLeft={() => {}}
-    />
-  );
+test("renders Timeline component", () => {
+  render(<Timeline timeLeft={30} />);
 
   const timerElement = screen.getByText(/30s/i);
   expect(timerElement).toBeInTheDocument();
