+++ components/ResultListPage/MainMenu/index.js
@@ -4,7 +4,7 @@ import Link from "next/link";
 export default function MainMenu({ currentRound }) {
   if (currentRound === 3) {
     return (
-      <StyledLink aria-label="Pfeil" href="/mainMenu">
+      <StyledLink aria-label="Pfeil" href="/main-menu">
         &larr;
       </StyledLink>
     );
