+++ package.json
@@ -18,7 +18,8 @@
     "react-dom": "18.2.0",
     "react-icons": "^4.8.0",
     "styled-components": "^5.3.8",
-    "use-local-storage-state": "^18.3.2"
+    "use-local-storage-state": "^18.3.2",
+    "ws": "^8.13.0"
   },
   "devDependencies": {
     "@svgr/webpack": "^6.5.1",
