+++ pages/index.js
@@ -1,4 +1,4 @@
-import SignUp from "./signUp";
+import SignUp from "./sign-up";
 
 export default function Starten({ nickname, setNickname }) {
   return <SignUp nickname={nickname} setNickname={setNickname} />;
